Refactor emploi page for web and io
